refactor(core): rename helpers folder by common
