+++ src/domain/common/store.helper.ts
