+++ src/domain/error/usecase/acknowledge-error/actionCreators.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-import type { ActionsUnion } from 'domain/helpers/store.helper'
-import { createAction } from 'domain/helpers/store.helper'
+import type { ActionsUnion } from 'domain/common/store.helper'
+import { createAction } from 'domain/common/store.helper'
 
 export const AcknowledgeErrorActions = {
   errorAcknowledged: () => createAction('error/errorAcknowledged')
