+++ src/domain/error/usecase/report-error/actionCreators.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import type { Error } from 'domain/error/entity/error'
-import type { ActionsUnion } from 'domain/helpers/store.helper'
-import { createAction } from 'domain/helpers/store.helper'
+import type { ActionsUnion } from 'domain/common/store.helper'
+import { createAction } from 'domain/common/store.helper'
 
 export const ReportErrorActions = {
   errorReported: (error: Error) => createAction('error/errorReported', { error })
