+++ src/domain/faucet/usecase/set-address/actionCreators.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-import type { ActionsUnion } from '../../../helpers/store.helper'
-import { createAction } from '../../../helpers/store.helper'
+import type { ActionsUnion } from 'domain/common/store.helper'
+import { createAction } from 'domain/common/store.helper'
 
 export const SetAddressActions = {
   addressSet: (address: string) => createAction('faucet/addressSet', { address })
