+++ src/domain/wallet/usecases/actionCreators.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-import type { ActionsUnion } from '../../helpers/store.helper'
-import { createAction } from '../../helpers/store.helper'
+import type { ActionsUnion } from 'domain/common/store.helper'
+import { createAction } from 'domain/common/store.helper'
 import type { ConnectionError, UnspecifiedError } from 'domain/wallet/entities/errors'
-import type { DeepReadonly } from '../../../superTypes'
+import type { DeepReadonly } from 'superTypes'
 
 export const ErrorWalletActions = {
   walletFailed: (error: DeepReadonly<ConnectionError> | DeepReadonly<UnspecifiedError>) =>
